Add files via upload
Spoilboard files are the current recommended work holding fixture for Shapeoko 3, it's a smaller cutting area, but should have no issues working with the SO3-XL and SO3-XXL. I don't have access to a <SO3. 

These boards are intended to go on top of the existing board that makes up the base of the Shapeoko. Once the holes are drilled and cased with an appropriate 9mm drill (I used an 11/32 drill..) you can put in standard threaded inserts available from most home improvement stores. Most of these seem to be 20mm, which would require purchasing a board that is at least 3/4" thick and then cutting to the size shown in the Spoilboard -Sheet1.pdf file. You absolutely could make this board without the use of the shapeoko, it's just labor intensive to do so. 

For ease of drawing, this board is technically rotated 90deg, so the long axis (22") will make up the Y axis of your ShapeOko, the last set of holes that can't be drilled due to the traverse limitation of the shapeoko can either be om